Trial: Split shared email method off to a concern
`can_email_user_question?`
