+++ app/controllers/emails_controller.rb
@@ -2,6 +2,8 @@
 
 # Send emails directly to webmaster and users via the application
 class EmailsController < ApplicationController
+  include ::Emailable
+
   before_action :login_required, except: [
     :ask_webmaster_question
   ]
@@ -51,16 +53,6 @@ class EmailsController < ApplicationController
     CommercialMailer.build(@user, @image, commercial_inquiry).deliver_now
     flash_notice(:runtime_commercial_inquiry_success.t)
     redirect_with_query(image_path(@image.id))
-  end
-
-  def can_email_user_question?(target, method: :email_general_question)
-    user = target.is_a?(User) ? target : target.user
-    return true if user.send(method) && !user.no_emails
-
-    flash_error(:permission_denied.t)
-    redirect_with_query(controller: target.show_controller,
-                        action: target.show_action, id: target.id)
-    false
   end
 
   def merge_request
